refactor(stacks): refactor creating of command context
affects: @takomo/cli, @takomo/config-repository-fs, @takomo/core, @takomo/organization-commands,
@takomo/test-integration, @takomo/util
