--- packages/cli/src/cli-command-context.ts
@@ -1,30 +1,0 @@
-import { InternalCommandContext } from "@takomo/core"
-import { FilePath } from "@takomo/util"
-
-export interface ProjectFilePaths {
-  readonly projectDir: FilePath
-  readonly stacksDir: FilePath
-  readonly templatesDir: FilePath
-  readonly partialsDir: FilePath
-  readonly hooksDir: FilePath
-  readonly resolversDir: FilePath
-  readonly helpersDir: FilePath
-  readonly schemasDir: FilePath
-  readonly projectConfigFile: FilePath
-  readonly stackGroupConfigFileName: string
-  readonly projectConfigFileName: string
-  readonly configFileExtension: string
-  readonly defaultDeploymentConfigFileName: string
-  readonly deploymentDir: FilePath
-  readonly configSetsDir: FilePath
-  readonly organizationDir: FilePath
-  readonly organizationTagPoliciesDir: FilePath
-  readonly organizationBackupPoliciesDir: FilePath
-  readonly organizationServiceControlPoliciesDir: FilePath
-  readonly organizationAiServicesOptOutPoliciesDir: FilePath
-  readonly defaultOrganizationConfigFileName: string
-}
-
-export interface CliCommandContext extends InternalCommandContext {
-  readonly filePaths: ProjectFilePaths
-}
