+++ packages/cli/test/options/parse-var-args.test.ts
@@ -1,32 +1,32 @@
-import { parseVarArgs } from "../src/options/parse-var-args"
+import { parseVarArgs } from "../../src/options/parse-var-args"
 
-describe("parse var args", () => {
-  test("returns an empty object when null is given", async () => {
+describe("#parseVarArgs", () => {
+  test("returns an empty object when null is given", () => {
     const vars = parseVarArgs(null)
     expect(vars).toStrictEqual({})
   })
 
-  test("returns an empty object when undefined is given", async () => {
+  test("returns an empty object when undefined is given", () => {
     const vars = parseVarArgs(undefined)
     expect(vars).toStrictEqual({})
   })
 
-  test("returns an empty object when empty array is given", async () => {
+  test("returns an empty object when empty array is given", () => {
     const vars = parseVarArgs([])
     expect(vars).toStrictEqual({})
   })
 
-  test("simple variable assignment", async () => {
+  test("simple variable assignment", () => {
     const vars = parseVarArgs(["name=john"])
     expect(vars).toStrictEqual({ name: "john" })
   })
 
-  test("multiple simple variable assignments", async () => {
+  test("multiple simple variable assignments", () => {
     const vars = parseVarArgs(["age=32", "code=secret"])
     expect(vars).toStrictEqual({ age: "32", code: "secret" })
   })
 
-  test("latest variable overrides variables defined before", async () => {
+  test("latest variable overrides variables defined before", () => {
     const vars = parseVarArgs(["age=32", "code=secret", "code=public"])
     expect(vars).toStrictEqual({ age: "32", code: "public" })
   })
