+++ packages/config-repository-fs/src/index.ts
@@ -1,3 +1,7 @@
+export {
+  createFileSystemCommandContext,
+  FileSystemCommandContext,
+} from "./context/create-file-system-command-context"
 export { createFileSystemDeploymentTargetsConfigRepository } from "./deployment-targets/config-repository"
 export { createFileSystemProjectConfigRepository } from "./init/config-repository"
 export { createFileSystemOrganizationConfigRepository } from "./organizations/config-repository"
