+++ packages/config-repository-fs/test/project-config-02.yml
