+++ packages/config-repository-fs/test/project-config-03.yml
