+++ packages/config-repository-fs/test/project-config-04.yml
