+++ packages/config-repository-fs/test/project-config-05.yml
