+++ packages/config-repository-fs/test/project-config-06.yml
