+++ packages/config-repository-fs/test/project-config-07.yml
