+++ packages/config-repository-fs/test/project-config-08.yml
